Change order of words in dictionary in test
If the high_typeability_words words are written to the dictionary first,
an easy and unintentional way of making the tests go green is to simply
read the first words in the dictionary.

While not a great improvement, this change in the test structure will
prevent the silliest of implementations from making the test go green.