Add CLI flag for choosing layout, Layout and tests
Make the argument parser and CommandLineOutput handle layouts.
The Layout class added is a skeleton class. Current intention is to make
the #left_hand_motions method work properly before a #right_hand_motions
method is added.

The tests for the #left_hand_motions method are mostly comprised of
sample sequences that are required. For everything but the one letter
sequences (and maybe the two letter sequences), a comprehensive test
would require too many sequences to write out, which is why arbitrary
samples are used instead.

The tests for three and four letter sequences are working but need
improvement.

The forbidden sequences for the two letter sequences might need to be
expanded if I figure out that certain sequences feel unacceptably bad to
type.