Add layout file reading capabilities to Layout
This mostly duplicates the file reading capabilities in Dictionary and
will therefore probably be extracted into dedicated file reading class
that can be shared in a future commit.

The qwerty.layout file was renamed to work with the copy-pasted code
from Dictionary, but after consideration I think it's fine to have no
file ending since it's placed in a dedicated directory.