Implement some left hand motions
This patch adds the complete generation of left hand one letter
sequences and the partial generation of two letter sequences.

The two letter sequences that are missing are those where the keys are
on adjacent rows but not on the same (or adjacent) columns. Example
sequences (on a QWERTY keyboard) that are missing would be AE, EG, FQ
and VA.

Also make tests for left hand three and four letter motions pending.
There's no implementation done for this and the tests are significantly
lacking in exhaustiveness.