Add Motion and Key classes
I found strings to be too primitive a construct to work as motions; same
thing for individual characters as representations of keys in a Layout.

Subsequently the classes Layout::Key and Motion were added.
Layout::Key is no more than a Struct that holds the character and its
position in the Layout.
Motion primarily houses sequences of Layout::Key and displays a string
representation of the sequences.

The currently unfinished left hand motion strategies needed to be
altered to move from working with nothing but characters, to working
with lists of Keys and returning Motions.

Also change references of "chars" to "keys".