Implement MotionGenerator
MotionGenerator replaces the current (rather manual) method of
generating motions. Instead of writing out the motions explicitly,
MotionGenerator generates all possible motions, where after it filters
away the motions that match against a set of forbidden ones.

While the filtering works, there's currently no good way to generate the
lists of forbidden motions. This will need to be addressed in future
commits.

In the tests for MotionGenerator, instances of Motion are created
instead of test doubles as it's usually done in the other tests. This is
because for some reason when doubles were passed in as restrictions,
they didn't respond to :keys messages, regardless of whether the double
was configured to expect such messages. It turned out to be easier to
just use Motion instances, so I used them in all relevant places.