Add MotionBanRule class
To be able to effectively filter Motions I wrote a rule DSL of sorts.
The idea is that upon instantiation of a MotionBanRule you get to supply
it with a block in which you're given a number of keywords that get
provided to you through the magic of instance_eval.

The first version of this made use of lambdas, which have the nice
property of allowing you to do a return inside of them (returning true
or false easily) which allowed one to just call them with a motion and
get the boolean response as to whether the motion was banned or not.

However, the lambda implementation had the problem of needing a bunch of
boilerplate in every rule. Additionally, accessing the "next key" was
a wordy and confusing mess.

This implementation makes use of catch-throw because it's the only way
I could think of that allows the user to return true or false in the
block in a straightforward way.