Unban two two-letter motions
After some consideration I decided that "qb" and "bq" are not
sufficiently unwieldy motions that they should be outright banned.

Later on they might get scored low (when the scoring algorithms are
implemented), but they shouldn't be banned.

Also re-flow the motions line to fill up the maximum line width.