Add interfaces necessary to make tests green
These are the remaining minor additions necessary to make the main
acceptance test for the "typeable" algorithm pass.

The fact that it passes is a problem in itself (there are some
constraints missing in the test), but at least these changes address the
issue of errors being raised when running the tests.