Move motions from Layout to MotionGenerator
Currently layout is both a kind of data object (a representation of
a keyboard layout stored in a text file) and something that knows about
how it can generate a bunch of motions based on itself.

This patch moves the motion generation to the MotionGenerator, which has
up until now been taking Key rows and amount of letters in the motions
it should generate. It now answers the question "For this instance of
Layout, what left hand and right hand motions can I get?".

Additionally, make the Layout#keys method public, to make Layout work
more like a collection of Keys. Flatten the Keys from Layout#keys, since
positional data is already embedded in each key, making a list of lists
resembling a keyboard redundant and needlessly complicated.