Extract pending test for duplicate passwords
Everything in the typeable password feature spec is green apart from the
test for duplicates. Implementing the code which ensures that there are
no duplicates is quite far off, since finishing MotionGenerator and the
Motion scoring algorithm are prioritized higher. Subsequently, to avoid
having the test for duplicates constantly be red, put it as pending.